fix build type issue
